requirements.txt
accelerate==0.17.1
einops==0.7.0
hydra-core==1.3.2
lpips==0.1.4
numpy==1.23.5
omegaconf==2.3.0
opencv_python_headless==4.7.0.72
packaging==23.2
Pillow==10.2.0
pytorch_lightning==1.6.3
scipy==1.10.1
scikit-image==0.21.0
torch==2.0.0
torchdiffeq==0.2.3
torchsde==0.2.5
torchvision==0.15.1
tqdm==4.65.0
